refactor(client): split TabbedLayout components into separate elements
