+++ packages/client/src/elements/layouts/TabbedLayout/Tab/Tab.styled.ts
@@ -1,47 +1,6 @@
 import styled from 'styled-components';
 
-import {
-  TabButtonProps,
-  TabLabelProps,
-  TabMarkerPosition,
-  TabsIndicatorProps,
-  TabsWrapperProps,
-} from './TabbedLayout.types';
-
-export const TabsWrapper = styled.div<TabsWrapperProps>`
-  position: relative;
-  overflow-x: auto;
-  display: flex;
-  height: 100%;
-  width: 100%;
-
-  ${({ variant }) => {
-    if (variant === 'fullWidth') {
-      return 'justify-content: space-evenly;';
-    }
-  }}
-`;
-
-const getTabsIndicatorPosition = ({ position }: { position: TabMarkerPosition }) => {
-  if (position === 'bottom') {
-    return 'bottom: 0';
-  }
-  if (position === 'top') {
-    return 'top: 0';
-  }
-};
-
-export const TabsIndicator = styled.span<TabsIndicatorProps>`
-  position: absolute;
-  bottom: 0;
-  ${getTabsIndicatorPosition};
-  background: ${({ theme }) => theme.palette.primary.light};
-  left: ${({ left }) => `${left}px`};
-  height: 5px;
-  width: ${({ width }) => width}px;
-  transition: left 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-  border-radius: ${({ theme }) => theme.sizes.borderRadius};
-`;
+import { TabButtonProps, TabLabelProps } from './Tab.types';
 
 export const TabLabel = styled.p<TabLabelProps>`
   font-size: 14px;
@@ -81,8 +40,4 @@ export const TabButton = styled.button<TabButtonProps>`
     box-shadow: inset 0 0 0 2px ${({ theme }) => theme.palette.primary.light};
     border-radius: ${({ theme }) => theme.sizes.borderRadius};
   }
-`;
-
-export const TabPanelWrapper = styled.div`
-  padding: 30px;
 `;
