+++ packages/client/src/elements/layouts/TabbedLayout/TabPanel/TabPanel.styled.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const TabPanelWrapper = styled.div`
+  padding: 30px;
+`;
