+++ packages/client/src/elements/layouts/TabbedLayout/TabPanel/TabPanel.types.d.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+interface TabPanelProps {
+  value: number;
+  index: number;
+  children: ReactNode;
+}
