+++ packages/client/src/elements/layouts/TabbedLayout/index.ts
@@ -1,0 +1,9 @@
+import Tab from './Tab';
+import TabPanel from './TabPanel';
+import Tabs from './Tabs';
+
+export default {
+  Tabs,
+  Tab,
+  TabPanel,
+};
